feat: rename SpectacularPipeHarness
